feat: affichage des items de la liste correspondante uniquement et amélioration de l'interface

+++ src/Controllers/ItemController.php
@@ -43,42 +43,43 @@ class ItemController extends BaseController
      */
     public function newItem(Request $request, Response $response, array $args): Response
     {
-        try {
-            $body = $request->getParsedBody();
-            $body = array_map(function ($field) {
-                return filter_var($field, FILTER_SANITIZE_STRING);
-            }, $body);
-
-            try {
-                Validator::failIfEmptyOrNull($body, ['image']);
-            } catch(Exception $e) {
-                Flashes::addFlash($e->getMessage(), 'error');
-                return $response->withRedirect($this->pathFor('newItemPage'));
-            }
-
-            $files = $request->getUploadedFiles();
-            $file = $files['image'];
-            if ($file->getError() === UPLOAD_ERR_OK) {
-                $directory = ROUTE . '\img\\';
-                $filename = UploadFile::moveUploadedFile($directory, $file);
-            } 
-
-            $item = new Item();
-            $item->list_id = $body['list_id'];
-            $item->name = $body['name'];
-            $item->description = $body['description'];
-            $item->image = $filename;
-            $item->url = $body['url'];
-            $item->price = $body['price'];
-            $item->save();
-
-            return $response->withRedirect($this->pathFor('displayAllItems'));
-        } catch (ModelNotFoundException $e) {
+        $list = WishList::find($args['list_id']);
+        if (!$list) {
             Flashes::addFlash('Impossible de créer l\'item', 'error');
-            return $response->withRedirect($this->pathFor('displayAllItems'));
+            return $response->withRedirect($this->pathFor('displayAllLists'));
         }
+
+        $body = $request->getParsedBody();
+        $body = array_map(function ($field) {
+            return filter_var($field, FILTER_SANITIZE_STRING);
+        }, $body);
+
+        try {
+            Validator::failIfEmptyOrNull($body, ['image', 'url']);
+        } catch (Exception $e) {
+            Flashes::addFlash($e->getMessage(), 'error');
+            return $response->withRedirect($this->pathFor('newItemPage', ['list_id' => $list->id]));
+        }
+
+        $files = $request->getUploadedFiles();
+        $file = $files['image'];
+        if ($file->getError() === UPLOAD_ERR_OK) {
+            $directory = ROUTE . '\img\\';
+            $filename = UploadFile::moveUploadedFile($directory, $file);
+        }
+
+        $item = new Item();
+        $item->list_id = $list->id;
+        $item->name = $body['name'];
+        $item->description = $body['description'];
+        $item->image = $filename;
+        $item->url = $body['url'];
+        $item->price = $body['price'];
+        $item->save();
+
+        return $response->withRedirect($this->pathFor('displayAllItems', ['list_id' => $list->id]));
     }
-    
+
     /**
      * Crée une vue pour afficher les items
      *
@@ -89,16 +90,18 @@ class ItemController extends BaseController
      */
     public function displayAllItems(Request $request, Response $response, array $args): Response
     {
-        try {
-            $items = Item::all();
-
-            $v = new ItemView($this->container, ['items' => $items]);
-            $response->getBody()->write($v->render(1));
-            return $response;
-        } catch (ModelNotFoundException $e) {
-            Flashes::addFlash("L'item " . $args['id'] . " n'a pas été trouvé", 'error');
-            return $response;
+        $list = WishList::with('items')->find($args['list_id']);
+        if (!$list) {
+            Flashes::addFlash("Liste introuvable.", 'error');
+            return $response->withRedirect($this->pathFor('displayAllLists'));
         }
+
+        $v = new ItemView($this->container, [
+            'list' => $list,
+            'items' => $list->items
+        ]);
+        $response->getBody()->write($v->render(1));
+        return $response;
     }
 
     /**
@@ -112,18 +115,17 @@ class ItemController extends BaseController
     public function displayItem(Request $request, Response $response, array $args): Response
     {
         $list = WishList::where('token', $args['token'])->first();
-
         if (!$list) {
             Flashes::addFlash("Liste introuvable.", 'error');
             return $response->withRedirect($this->pathFor('home'));
         }
-        
-        $item = Item::find($args['id']);
+
+        $item = Item::find($args['item_id']);
         if (!$item) {
             Flashes::addFlash("Item introuvable.", 'error');
             return $response->withRedirect($this->pathFor('home'));
         }
-        
+
         if ($list->id !== $item->list_id) {
             Flashes::addFlash("L'item ne correspond pas à la liste.", 'error');
             return $response->withRedirect($this->pathFor('home'));
@@ -147,16 +149,21 @@ class ItemController extends BaseController
      */
     public function editItemPage(Request $request, Response $response, array $args): Response
     {
-        try {
-            $item = Item::findOrFail($args['id']);
-
-            $v = new ItemView($this->container, ['item' => $item]);
-            $response->getBody()->write($v->render(3));
-            return $response;
-        } catch (ModelNotFoundException $e) {
-            Flashes::addFlash("L'item " . $args['id'] . " n'a pas été trouvé", 'error');
-            return $response->withRedirect($this->pathFor('displayAllItems'));
+        $list = WishList::find($args['list_id']);
+        if (!$list) {
+            Flashes::addFlash("Liste introuvable.", 'error');
+            return $response->withRedirect($this->pathFor('displayAllLists'));
         }
+
+        $item = Item::find($args['item_id']);
+        if (!$item || $list->id !== $item->list_id) {
+            Flashes::addFlash("L'item n'a pas été trouvé", 'error');
+            return $response->withRedirect($this->pathFor('displayAllItems', ['list_id' => $list->id]));
+        }
+
+        $v = new ItemView($this->container, ['list' => $list, 'item' => $item]);
+        $response->getBody()->write($v->render(3));
+        return $response;
     }
 
     /**
@@ -169,47 +176,54 @@ class ItemController extends BaseController
      */
     public function editItem(Request $request, Response $response, array $args): Response
     {
-        try {
-            $item = Item::findOrFail($args['id']);
-
-            $body = $request->getParsedBody();
-            $body = array_map(function ($field) {
-                return filter_var($field, FILTER_SANITIZE_STRING);
-            }, $body);
-
-            try {
-                Validator::failIfEmptyOrNull($body, ['image', 'url']);
-            } catch(Exception $e) {
-                Flashes::addFlash($e->getMessage(), 'error');
-                return $response->withRedirect($this->pathFor('editItemPage', ['id' => $args['id']]));
-            }
-            
-            $files = $request->getUploadedFiles();
-            $file = $files['image'];
-            if($file !== null) {
-                if ($file->getError() === UPLOAD_ERR_OK) {
-                    $directory = ROUTE . '\img\\';
-                    unlink($directory . $item->image);
-                    $filename = UploadFile::moveUploadedFile($directory, $file);
-                }
-            } else {
-                $filename = $item->image;
-            }
-
-            $item->list_id = $body['list_id'];
-            $item->name = $body['name'];
-            $item->description = $body['description'];
-            $item->image = $filename;
-            $item->url = $body['url'];
-            $item->price = $body['price'];
-            $item->save();
-
-            Flashes::addFlash("Item modifié avec succès", 'success');
-            return $response->withRedirect($this->pathFor('displayAllItems'));
-        } catch (Throwable $e) {
-            Flashes::addFlash("Impossible de modifier l'item", 'error');
-            return $response->withRedirect($this->pathFor('displayAllItems'));
+        $list = WishList::find($args['list_id']);
+        if (!$list) {
+            Flashes::addFlash("Liste introuvable.", 'error');
+            return $response->withRedirect($this->pathFor('displayAllLists'));
         }
+
+        $item = Item::find($args['item_id']);
+        if (!$item || $list->id !== $item->list_id) {
+            Flashes::addFlash("L'item n'a pas été trouvé", 'error');
+            return $response->withRedirect($this->pathFor('displayAllItems', ['list_id' => $list->id]));
+        }
+
+        $body = $request->getParsedBody();
+        $body = array_map(function ($field) {
+            return filter_var($field, FILTER_SANITIZE_STRING);
+        }, $body);
+
+        try {
+            Validator::failIfEmptyOrNull($body, ['image', 'url']);
+        } catch (Exception $e) {
+            Flashes::addFlash($e->getMessage(), 'error');
+            return $response->withRedirect($this->pathFor('editItemPage', [
+                'list_id' => $list->id,
+                'item_id' => $item->id
+            ]));
+        }
+
+        $files = $request->getUploadedFiles();
+        $file = $files['image'];
+        if ($file !== null) {
+            if ($file->getError() === UPLOAD_ERR_OK) {
+                $directory = ROUTE . '\img\\';
+                unlink($directory . $item->image);
+                $filename = UploadFile::moveUploadedFile($directory, $file);
+            }
+        } else {
+            $filename = $item->image;
+        }
+
+        $item->name = $body['name'];
+        $item->description = $body['description'];
+        $item->image = $filename;
+        $item->url = $body['url'];
+        $item->price = $body['price'];
+        $item->save();
+
+        Flashes::addFlash("Item modifié avec succès", 'success');
+        return $response->withRedirect($this->pathFor('displayAllItems', ['list_id' => $list->id]));
     }
 
     /**
@@ -222,22 +236,28 @@ class ItemController extends BaseController
      */
     public function deleteItem(Request $request, Response $response, array $args): Response
     {
-        try {
-            $item = Item::findOrFail($args['id']);
-            $directory = ROUTE . '\img\\';
-            unlink($directory . $item->image);
-            $item->delete();
-
-            Flashes::addFlash("Item supprimé avec succès", 'success');
-            return $response->withRedirect($this->pathFor('displayAllItems'));
-        } catch (ModelNotFoundException $e) {
-            Flashes::addFlash("Impossible de supprimer l'item", 'error');
-            return $response->withRedirect($this->pathFor('displayAllItems'));
+        $list = WishList::find($args['list_id']);
+        if (!$list) {
+            Flashes::addFlash("Liste introuvable.", 'error');
+            return $response->withRedirect($this->pathFor('displayAllLists'));
         }
+
+        $item = Item::find($args['item_id']);
+        if (!$item || $list->id !== $item->list_id) {
+            Flashes::addFlash("L'item n'a pas été trouvé", 'error');
+            return $response->withRedirect($this->pathFor('displayAllItems', ['list_id' => $list->id]));
+        }
+
+        $directory = ROUTE . '\img\\';
+        unlink($directory . $item->image);
+        $item->delete();
+
+        Flashes::addFlash("Item supprimé avec succès", 'success');
+        return $response->withRedirect($this->pathFor('displayAllItems', ['list_id' => $list->id]));
     }
 
-     /**
-     * formulaire de réservation d'un item
+    /**
+     * Formulaire de réservation d'un item
      * 
      * @param Request $request requête
      * @param Response $response réponse
@@ -246,19 +266,24 @@ class ItemController extends BaseController
      */
     public function lockItemPage(Request $request, Response $response, array $args): Response
     {
-        try {
-            $item = Item::with('list')->findOrFail($args['id']);
-
-            $v = new ItemView($this->container, [
-                'item' => $item,
-                'list' => $item->list
-            ]);
-            $response->getBody()->write($v->render(4));
-            return $response;
-        } catch (ModelNotFoundException $e) {
-            Flashes::addFlash("L'item " . $args['id'] . " n'a pas été trouvé", 'error');
-            return $response;
+        $list = WishList::find($args['list_id']);
+        if (!$list) {
+            Flashes::addFlash("Liste introuvable.", 'error');
+            return $response->withRedirect($this->pathFor('home'));
         }
+
+        $item = Item::find($args['item_id']);
+        if (!$item || $list->id !== $item->list_id) {
+            Flashes::addFlash("L'item n'a pas été trouvé", 'error');
+            return $response->withRedirect($this->pathFor('displayList', ['list_id' => $list->id]));
+        }
+
+        $v = new ItemView($this->container, [
+            'item' => $item,
+            'list' => $item->list
+        ]);
+        $response->getBody()->write($v->render(4));
+        return $response;
     }
 
     /**
@@ -271,36 +296,41 @@ class ItemController extends BaseController
      */
     public function lockItem(Request $request, Response $response, array $args): Response
     {
-        try {
-            $item = Item::with('reservation')->findOrFail($args['id']);
-
-            if (!$item->reservation) {
-                $user = Auth::getUser();
-
-                $body = $request->getParsedBody();
-                $message = filter_var($body['message'], FILTER_SANITIZE_STRING);
-                
-                $reservation = new ItemReservation();
-                $reservation->item_id = $item->id;
-                $reservation->user_id = $user['id'];
-                $reservation->message = $message === '' ? null : $message;
-                $reservation->save();
-
-                Flashes::addFlash("Item réservé avec succès.", 'success');
-            } else {
-                Flashes::addFlash("Item déjà réservé.", 'error');
-            }
-            
-            $redirectUrl = $this->pathFor('displayItem', [
-                'token' => $item->list->token,
-                'id' => $item->id
-            ]);
-            return $response->withRedirect($redirectUrl);
-        } catch (\Throwable $th) {
-            Flashes::addFlash("Impossible de réservé l'item", 'error');
-            return $response->withRedirect($this->pathFor('displayAllItems'));
+        $list = WishList::find($args['list_id']);
+        if (!$list) {
+            Flashes::addFlash("Liste introuvable.", 'error');
+            return $response->withRedirect($this->pathFor('home'));
         }
-    }
+
+        $item = Item::find($args['item_id']);
+        if (!$item || $list->id !== $item->list_id) {
+            Flashes::addFlash("L'item n'a pas été trouvé", 'error');
+            return $response->withRedirect($this->pathFor('displayList', ['list_id' => $list->id]));
+        }
+
+        if (!$item->reservation) {
+            $user = Auth::getUser();
+
+            $body = $request->getParsedBody();
+            $message = filter_var($body['message'], FILTER_SANITIZE_STRING);
+
+            $reservation = new ItemReservation();
+            $reservation->item_id = $item->id;
+            $reservation->user_id = $user['id'];
+            $reservation->message = $message === '' ? null : $message;
+            $reservation->save();
+
+            Flashes::addFlash("Item réservé avec succès.", 'success');
+        } else {
+            Flashes::addFlash("Item déjà réservé.", 'error');
+        }
+
+        $redirectUrl = $this->pathFor('displayItem', [
+            'token' => $item->list->token,
+            'item_id' => $item->id
+        ]);
+        return $response->withRedirect($redirectUrl);
+}
 
     /**
      * Annulation de la réservation d'un item
@@ -312,22 +342,27 @@ class ItemController extends BaseController
      */
     public function cancelLockItem(Request $request, Response $response, array $args): Response
     {
-        try {
-            $item = Item::findOrFail($args['id']);
-           
-            $lock_message = ItemReservation::select('*')->where('item_id', '=', $item->reservation->item_id)->firstOrFail();
-            $lock_message->delete();
-              
-            Flashes::addFlash("Réservation annulée avec succès", 'success');
-
-            $redirectUrl = $this->pathFor('displayItem', [
-                'token' => $item->list->token,
-                'id' => $item->id
-            ]);
-            return $response->withRedirect($redirectUrl);
-        } catch (ModelNotFoundException $e) {
-            Flashes::addFlash("Impossible d'annuler la réservation de l'item", 'error');
-            return $response->withRedirect($this->pathFor('displayAllItems'));
+        $list = WishList::find($args['list_id']);
+        if (!$list) {
+            Flashes::addFlash("Liste introuvable.", 'error');
+            return $response->withRedirect($this->pathFor('home'));
         }
+
+        $item = Item::find($args['item_id']);
+        if (!$item || $list->id !== $item->list_id) {
+            Flashes::addFlash("L'item n'a pas été trouvé", 'error');
+            return $response->withRedirect($this->pathFor('displayList', ['list_id' => $list->id]));
+        }
+
+        $lock_message = ItemReservation::select('*')->where('item_id', $item->reservation->item_id)->firstOrFail();
+        $lock_message->delete();
+
+        Flashes::addFlash("Réservation annulée avec succès", 'success');
+
+        $redirectUrl = $this->pathFor('displayItem', [
+            'token' => $item->list->token,
+            'item_id' => $item->id
+        ]);
+        return $response->withRedirect($redirectUrl);
     }
 }
